SimpleMinigames/requirements.txt
pygame
Pillow
nltk